// web/back-office/src/app/api/model/edit-item-stat-request.ts
/**
 * Pirate\'s Odyssey - API
 * Swagger for Pirate\'s Odyssey API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Stats } from './stats';


export interface EditItemStatRequest { 
    id?: string;
    stats?: Stats;
    value?: number;
    itemId?: string;
}
export namespace EditItemStatRequest {
}